[Feat] - 로그인/로그아웃 시 FCM 토큰 처리

+++ src/main/java/com/ssu/ongi/domain/auth/service/AuthCommandService.java
@@ -1,7 +1,9 @@
 package com.ssu.ongi.domain.auth.service;
 
+import com.ssu.ongi.common.exception.GeneralException;
 import com.ssu.ongi.common.jwt.TokenCommandService;
 import com.ssu.ongi.common.jwt.TokenPair;
+import com.ssu.ongi.common.status.ErrorStatus;
 import com.ssu.ongi.domain.auth.dto.response.ReissueResponse;
 import com.ssu.ongi.domain.elder.entity.Elder;
 import com.ssu.ongi.domain.elder.service.ElderCommandService;
@@ -11,6 +13,7 @@ import com.ssu.ongi.domain.member.dto.request.SignupRequest;
 import com.ssu.ongi.domain.member.dto.request.UpdatePasswordRequest;
 import com.ssu.ongi.domain.auth.dto.response.LoginResponse;
 import com.ssu.ongi.domain.member.entity.Member;
+import com.ssu.ongi.domain.member.enums.LoginMode;
 import com.ssu.ongi.domain.member.service.MemberCommandService;
 import com.ssu.ongi.domain.member.service.MemberQueryService;
 import lombok.RequiredArgsConstructor;
@@ -41,13 +44,26 @@ public class AuthCommandService {
         Member member = memberQueryService.findByLoginIdWithElders(request.loginId());
         memberQueryService.validatePassword(member, request.password());
 
+        if (request.loginMode() == LoginMode.GUARDIAN) {
+            memberCommandService.updateFcmToken(member, request.fcmToken(), request.osType());
+        } else {
+            Elder elder = member.getElders().stream()
+                    .findFirst()
+                    .orElseThrow(() -> new GeneralException(ErrorStatus.ELDER_NOT_FOUND));
+            elderCommandService.updateFcmToken(elder, request.fcmToken(),request.osType());
+        }
+
         TokenPair tokens = tokenCommandService.issueTokens(member.getId(), request.loginMode());
         return LoginResponse.of(tokens.accessToken(), tokens.refreshToken(), request.loginMode(), member);
     }
 
-    public void logout(Long memberId) {
+    public void logout(Long memberId, LoginMode loginMode) {
+        if (loginMode == LoginMode.ELDER) {
+            throw new GeneralException(ErrorStatus.ELDER_CANNOT_LOGOUT);
+        }
         tokenCommandService.logout(memberId);
         memberCommandService.deleteFcmToken(memberId);
+        elderCommandService.deleteFcmToken(memberId);
     }
 
     public void withdraw(Long memberId) {
